Change registry interface to have separate login, add push command (#88)
This changes the registry interface significantly as we move to having the login phase as a separate command (bay registry login). It also makes the registry plugin its own class with methods rather than a callable.

It also adds a bay push command and the related code in the images module to support it.