Fixed the bay doctor command
Summary: Fixed the bay doctor command. The doctor exam when doesn't fail return None, but the doctor validation set problem as True for None values. This validation was fixed

Test Plan:
Do an arc patch from this diff
Put the debugger in the file doctor.py on the line 101
Check if the exams result
The problem variable should be updated to True just if the exam_result is different then None

Reviewers: #sys_platform, remy

Reviewed By: #sys_platform, remy

Subscribers: michaelmanganiello

JIRA Issues: EB-68244

Differential Revision: https://phabricator.evbhome.com/D45524