Remove the volume "cached" suffix
developers sometimes have an issue where file changes on the mac are not picked up by the container, wondering if it's caused by the "cached" volume flag, trying to go back to the default of strong consistency to see if it fixes the issue (at the probable cost of lower performance). Just a test for now.

see https://docs.docker.com/docker-for-mac/osxfs-caching/#tuning-with-consistent-cached-and-delegated-configurations