Allow non file resources to exist on a repository
Summary:
If a user wants to add a resource that is not a file or a directory, the build
process will fail. That is something that is not necessary good as that resource
can be necessary outside of the container only.

Our particular case happened when we wanted to add symlinks to the .git/hooks
folder of payments and the build process started failing. Specifically, we wanted
to add a symlink from a folder that exists inside payment_service to the .git/hooks
folder in the same repository. After creating it, the build process would fail because
the link is neither a folder or a file.

With this change, the symlink is ignored when creating the tarball that will later be
used when doing the actual build

Test Plan:
- Add a non file resource to the repository, e.g. a symlink
- Rebuild the container
-- Excpect: Build process completes successfully

Reviewers: guillaume.carre, andrewgodwin, michaelmanganiello, cristian, #payments, remy

Reviewed By: michaelmanganiello, cristian, #payments, remy

Subscribers: michaelmanganiello, cristian

Differential Revision: https://phabricator.evbhome.com/D59641