EB-84956 when building a profile with the 'inherits' property, include the parent containers in the build
Summary:
Currently if a profile defined 'ignore-dependencies: true' and 'inherits: base', when doing `bay build`, it only builds the containers listed in the profile file, not the parent (in this example the 'base' profile). The workaround for the user is to switch profile, build and re-switch profile.
With this fix, the build will include the container listed in the ancestor profile as well, so the user doesn't need to switch profile when building.

Test Plan:
tested with the profile base-cerberus

Reviewers:

Subscribers:

JIRA Issues:
EB-84956