Use list format for binds instead of dict format for multiple mountpoints
Docker's `binds` argument (defined here as `volume_binds`) can be in two formats. It can be in a list of strings `'{source}:{destination}:{mode}'`, or it can be a dict whose keys are sources and whose values are a dict of `{'bind': '{destination}', 'mode': '{mode}'}`. If you specify `binds` in dict format, the Docker SDK converts it to list format before sending it to the Docker process. However, the dict format limits you to one container mountpoint per host source. Docker permits multiple container mountpoints per host source, and the only way to specify that is with the list format. Previously we used the dict format, but this change switches to the list format to support multiple mountpoints.